Test package further compatibility for CLI
The RML file as indicated by RML_Modules must exist in `src/mappings`.

WARNING: The CLI tools have a bug whereby the rml_module_injector
deletes the RML file in the package if the CM header for RML_Modules is
not right OR the file does not exist in `src/mappings`.