Add Buyer role, conditional instantiation attempt
CompetitionNotice should be declared through a complex XPath condition,
and so should the `announcesRole` property (only valid for such a
notice). This, however, does not seem to be possible through an RML
`reference` as it must refer to an element (and the condition deals with
several disparate elements), but it works if an entirely new TriplesMap
is created with the condition embedded in the iterator, where the
wildcard root can be set if true otherwise null (which essentially means
no iterator and therefore no results).