Add tentative language mapping construct
There is no direct way to encode language as the codes used are not
standard. This may require use of convoluated conditions or custom RML
functions. While we are investigating, we can have the language codes
transformed verbatim, which can be post-processed by a simple script to
replace the tags (which itself is another viable solution).

Also start mapping predicates using relative XPaths.