Add ContractTerm natures; potentially problematic
Previously feedback was received that `foreseesContractTerm` "appears
twice", and that "instantiations should be for the Nature of the
Contract and the Additional Nature of the contract – the Additional
Nature is missing".

While the additional nature does appear now in the output, there are
still two ContractTerms with the same main and additional contract
nature values, owing to two realized locations. It may be that the
additional needs to go in the additional realized location, but that's
just a guess.