Trial true language support w/ RML built-in function
As we need to transform input language data to BCP47 (`en` not `eng`)
for the language tags, we need to resort to RML functions. We were made
aware, by the author, of the built-in function `lookup`, saving us from
writing our own in Java.

Care must be taken to ensure that the namespace of the function is
correct for the version of RMLMapper (check with `-v` and search the
debug output for "Parsing new function"). For v6.2.2, the following
namespace will be different compared to official documentation:

```
@prefix idlab-fn: <http://example.com/idlab/function/> .
```

See https://rml.io/docs/rmlmapper/default-functions/#lookup