Remove non-working country mapping integration
Mapping to a parent vocabulary resource serves no purpose if it is not
expected to work and instead will create erroneous output (all possible
values).

This change also introduces change (side-effect) in output of RMLMapper
regardless of the version used, repositioning the `a cccev:ContactPoint`
type assertion to come on top of the block of related statements. It is
unknown why this happens particularly for this change, but it is a
desired outcome nonetheless (type assertions in the middle are not
expected although may very well be possible and valid).