Move out minimal test folder into src, turn into reference
The minimal test folder for direct RMLMapper testing alongside other
packages can be misleading. Furthermore, we would like a single source
of truth for the RML rules under development. The `src` folder already
serves that role and therefore can be repurposed, albeit with some
restructuring and additional documentation for clarity.