Use new lang vocab/technique to map at-voc:language
Repurpose the CSV voc used for language tags (`languageMap`) to map to
`at-voc:language` URIs, using the same technique of looking up against
the URI in order to get the URI. This is because only the URIs have
complete ISO coverage for the possible inputs. If we don't find a match
against the URIs, we simply don't map any value.