Add Procedure hasLegalBasis at-voc:legal-basis
We add the predicate mapping under the existing TMap of Procedure
anchored to the TenderingProcess node, as that is where it was declared
previously along with BT-01(c|d).

However, this is now possibly distinct, and it may be that we should
instead instantiate another TMap for Procedure anchored to ND-Root. On
the other hand, if several TMaps exist for one subject, it does not
matter where the predicate mapping is declared, especially if it's just
one of these whose XPath is orthogonal to its siblings.