Use joinCondition for OPT-300-Procedure-SProvider
Any joinCondition to a parent where the subject template includes an
explicit use of values must also point to values instead of paths. So,
for example, parents such as Organization whose IRIs include an internal
ID of the form `ORG-XXX`, vs. other resources whose IRIs are simply a
hash of the path.