Combine Procedure Legal Basis ID w/ Description
According to latest feedback ePO "does not foresee an identifier for the
local legal basis" and so we combine the ID with the description, along
with any `schemeName` although that never happens to exist in the data.

Additionally, fix some stylistic/cosmetic aspects, edit some comments,
remove some commented code made irrelevant by this change, and add some
version declarations.