Remove wrong instantiations related to Procedure
- Remove unnecessary Identifier mappings (owing to a misunderstanding or
unclear representation of linkages in the conceptual mapping)

- Remove wrong Lot, Procedure mappings (these create additional
instances; the correct number of instances for these already exist, i.e.
a single Procedure of the Contract, and two Lots)

- Remove wrong ProcedureTerm mapping owing to a wrong anchoring
(subject), made complicated by its absent parent, i.e. the LotGroup
(this is a Module 4 resource and, while it is separate from the one to
be instantiated for the ProcedureTerm, will also need to be instantiated
together with its Identifier so that assertions continue to work)

- Remove LotGroup instantiation to revisit later once the context is
clear (right now we don't have the main LotGroup as it is not part of
Module 2)