Fix Notice refersToPrevious external resources linking
The notice that is being referred to (the external resource) by the
current notice is under a child element that repeats. This has two major
implications:

1. The external notice being instantiated must include its ID in the
Notice part of the IRI instead of the current notice's, and in general
must conform to the IRI template of a Notice, which is a root
element/resource and therefore has a template different from the rest.

2. The linked child resources of this notice (the externally linked
resources; its Identifer, PlannedProcurementPart and the Identifier of
the latter) must conform to the IRI template of their parent and not of
the current notice, and additionally, must include a _hash component
that is done on an explicit element_ (the ID) rather than a path. This
is because hashing on a path would create additional instances as the
iterator is repeating.

This change also fixes a mistaken reference in a predicate mapping to
the notice ID when it should be another element (the internal address),
and removes unneeded commented out code (which look like leftover
duplicates).