Add ProcedureTerm, LotGroup, Lot relationships
Add mappings for BT-330-Procedure (ProcedureTerm -> LotGroup) and
BT-1375-Procedure (LotGroup -> Lot).

This is a complex situation involving subjects that have relationships
to objects at different levels, and therefore mapping groups use or
naming of triplesMaps is not consistent with the conceptual mapping.

Subjects sometimes need to be instantiated at some level above an
iterator, so as to be able to instantiate predicates/objects or
relationships for repeating elements.