Improve Makefile test target, add mixed types test
We want to be able to capture potential problems in subject
templates/references, where subjects that ought to be different
resources have the same IRIs. So, we query for any resource that has
more than one type declared. It's a blanket query but will give us a
clue, especially if we filter out the Notice, because that's the only
thing that may always have more than one type asserted.

Also change wording from "validate" to "test" (we in fact only _check_
but the testing terminology suffices for now).