Fix empty fullAddress in Procedure, rmlmapper v6.2.2 only
The Location/Address instance/relationship is unwarranted, as the
required elements are not present in the broad place of performance of
the procedure, which overlaps w/ the specific place of performance due
to the same XPath element `Address`.

Unfortunately, this does not work w/ RMLMapper v6.3+, likely due to the
null case and how the joinCondition works, because without a
joinCondition and the conditional reference with null, it does work.
Whether or not this is a regression needs to be consulted upstream.

Closes gh-22