Add missing rel for ProcedureTerm, rmlmapper v6.2.2 only
The `isSubmissionForAllLotsRequired` needs also be linked via the
`isSubjectToProcedureSpecificTerm` from the TriplesMap of the Procedure
at the same level `ND-ProcedureTenderingProcess`.

This does not work w/ RMLMapper v6.3+ due to the presence of a
joinCondition and null-conditioned reference where the data being tested
fails the condition but triggers a null error.

See https://github.com/RMLio/rmlmapper-java/issues/236