Fix null error from languageMap due to combination
Mapping combined fields where values are optional needs to be done
carefully, otherwise any missing value will trigger the languageMap
function to yield a null and therefore error out.

This was revealed when testing with the full suite of current sample
data, namely from the notice `725842-2023`, which has a proper legal
basis ID but not a description.

Using two objectMaps conditioned on the existence and lack of the
description, along with a fallback chain on the other two fields, helps
to nicely deal with all of the cases.

Also closes gh-26.