Do not instantiate ReviewTerm without any context
This instantiation is based on the CM and associated MG, however they do
not appear to be serving any purpose in the TM, as they created orphan
instances. These are most likely composition relationships only, meaning
there has to be a relationship and parent context for this resource to
exist.