Fix typos in IRI name component, Lot module 3
WARNING: Discovered that the resource having name
`SelectionTermConstraint` is not instantiated in singular test. Part of
BT-5423-Lot, requiring extension of that test data (source.xml).
However, this is not critical as the full test suite is likely to have
the necessary data and validation results will expose any error in the
mapping in question.