Role contact point implementation attempt, caveat emptor
It appears the XPath lookup does not work with relative XPaths embedded
inside the XPath filter conditions ("predicates"). That that would work
was perhaps a mistaken assumption during the design phase. It is not in
fact supported by XPath.

The link in a role to an organization via `epo:playedBy`, where the role
has a link to a touchpoint, will either be wrong or not exist, if the
data has _touchpoints in more than one organization_ (the RML reference
would find more than one element and therefore not work).

We might have to go back to the drawing board, and resort to things like
custom functions, in order to inject the relative XPath value into an
XPath query. OR, we can delegate this to a post-processing script, since
the `hasContactPointInRole` will exist as an indicator.

This affects all OPT-301-* fields, which are roles that _may_ link to a
touchpoint instead of an organization. We will only get lucky if data in
practice (and in clients' tests) does not frequently have that
situation.

See https://stackoverflow.com/questions/64285927/using-paths-in-xpath-predicates