Implement sameAs technique for hasContactPointInRole playedBy
- Instantiate a placeholder Organization w/ the ContactPoint ID
- Link the placeholder entity to the actual Organization
- Create a deterministic IRI at the role to link to the placeholder
- Rely on the `owl:sameAs` link and possible inference for materialization

The test and post-processing script for this will hopefully not yield
anything anymore as there will now be an `epo:playedBy` link where an
`epo:AgentInRole` (role) links to an `cpov:ContactPoint` (touchpoint).

This is possible because there is one Company element that is a sibling
to all TouchPoint elements under the Organization element. It will _not_
work if `efac:Company` ever becomes repeatable.

Co-authored-by: csnyulas <csongor.nyulas@gmail.com>