Fix BT-18-Lot incorrect datatype IRI vs. anyURI
User-supplied input URIs are _not_ the same as RDF URIs/IRIs, as they
are string-quoted data referring to URIs on the web, not RDF resources.
For this, we already have the practice of encoding them as `xsd:anyURI`,
which is seen throughout the mappings.

Issue exposed when testing against `725842-2023.xml`, which had an
EndpointID without the leading protocol segment (`http://`), which is
fine as a web URI, but not an RDF one.