Fix hasPrizeRank multiple values, unmap BT-45-Lot
The template path for the Prize subject was incorrect (at the wrong
level), and therefore its predicates had repeated values. op#16

As an indirectly related issue that has not yet been reported or
discovered: BT-45-Lot creates a discrepancy as descriptions are in the
prizes, which ePO does not foresee, and so the current mapping to the
DesignContestRegimeTerm's hasFollowupContractInformation is not
correct.