Initial versioned mappings, abs and rel XPath changes
Mappings of the following fields are the lucky first to be split up into
versioned RML files:

- BT-772-Lot only abs XPath
- BT-771-Lot both abs and rel Xpaths
- BT-729-Lot only abs XPath
- BT-64-Lot only abs Xpath
- BT-65-Lot unchanged, duplicated across all versions

It has been agreed that version-specific RML rules go to
version-specific RML files with a version range, while unversioned rules
remain in a "common" or shared file.

The following is an example file naming scheme for this approach, where
a `+` indicates no (or unknown) eForms SDK upper version limit, and
`0.0` no (or unknown) eForms SDK lower version limit:

- Lot.rml.ttl
- Lot_v1.3-1.6.rml.ttl
- Lot_v1.7+.rml.ttl
- Lot_v0.0-1.3.rml.ttl
- Lot_v1.5-1.8.rml.ttl

The versioned files themselves are put in another directory,
`mappings-versioned`, so that versioned folders and packages can be
derived, and RMLMapper can be run with multiple input arguments, for
example:

```sh
rmlmapper -m mappings/* mappings-1.9/* -s turtle
```

A development Bash shell script `prep-multiver.sh` is provided to create
the necessary versioned folders and populate them based on the version
ranges of the versioned RML files. These folders are not committed as
they will likely contain redundant files (a file with version
restriction v1.3-1.6 will be in all the versions in the range), and
should later serve as a basis for versioned packages.

Versioned output files are generated from transforming multiple versions
_but against the same reference input data_. These are committed in a
dedicated directory `output-versioned` so as to track how the migration
to versioned files affects the output. Note, however, save for the 1.9
output, the rest of the results may be incorrect as the input
`source.xml` is a specific version (modified cn_24_maximal 1.9).

When an RML mapping rule (a TriplesMap) contains predicate mappings
(predicateObjectMaps) of fields that have a significant change across
versions (XPath), most often the entire rule has to be migrated to a
versioned file, and common mappings part of the same rule cannot be kept
in the common file. This is because RMLMapper will raise violations for
duplicate subjectMaps, and has undetermined behaviour for duplicate
logicalSources.