Add and track multi-version SDK example and output
The reference data `source.xml` is a modified v1.9 cn_24_maximal.xml,
and will only serve to inform changes for that version. It will be
challenging to maintain several versions of the reference data with
their own modifications for coverage, so it is only appropriate to
introduce the reference example unchanged for all the versions, and keep
the output of those tracked alongside. This only has local development
relevance.