Add initial base versioned packages
Remove metadata until re-exported from validation tool (Mapping
Workbench).

Conceptual mapping file should be replaced with the specific filtered
versions, possibly in a Make package target (from `src` assuming we
start tracking those there) if not directly committed in these packages.

These are in fact "minimal" packages but the minimal suffix is expected
to be stripped, as it will be the default going forward.