Full support for versioned packaging, sans testing
- Adapt Makefile with support for looping over versions
- Rename random samples' subfolders to be consistent
- Sync has been separated (now must always be run before packaging)
- Versioned testing and analyses will have to be adapted similarly
- Some stray file executable permissions (on *NIX) are fixed

For a complete run, to create packages with ALL data wherever possible
(some won't have samples):

```sh
make clean && make && make package_cn_maximal && make export_cn_maximal
```

It is better to always clean up as there is currently a problem with
repeated runs messing up the output packages. To reiterate, sync (`make`
alone) must always be run now if intending to package.

It is no longer recommended to use the `*all_variants` target as it will
create 30+ packages. At most, one may create both the minimal and
maximal, and others as and when the need for them arises.

Some samples packages will be invalid (namely v1.4-1.5), as they do not
have samples, and currently the Makefile is not smart enough to skip
them from packaging or remove them after-the-fact.

Note that the Makefile features used may only be GNU-Make compatible.