Exclude again a problematic sample file from packaging, v1.6
It looks like there is an odd RMLMapper lookup function error again,
which was seen in the last file that was excluded. The function receives
null from something among the languageMaps, but it is next to impossible
to figure out where/what.

The transformation does not fail, but there is an error which usually
stops any parent process (like a pipeline or app running RMLMapper) as
there is stderr output from the tool.