LotProcurementDocument-related versioned mappings
This is perhaps the most unwieldy set of differences and corresponding
implementations. Several fields sharing the same TriplesMap and
sometimes parentTriplesMaps or the node have changes in different ways
in different version ranges. Affects the following fields:

- BT-737-Lot
- BT-707-Lot
- BT-708-Lot
- BT-15-Lot

WARNING: Versioned mappings still produce empty instances (runts), but
this is getting very difficult to debug and therefore investigation is
deferred.