Add support for releasing to client repository
The default client project is assumed to be a local directory sibling
relative to this project. However, this can be overridden by providing
the `RELEASE_DIR` variable as an argument to Make, e.g. `make
RELEASE_DIR=client-dir ...`.

WARNING: The SHACL file is not touched for now. It was getting annoying
trying to write the Bash `if` syntax in Make-compatible way (because the
Make `ifeq` evaluates only statically and so the dynamic variable
evaluates empty). Until next time..