Only CAN-relevant files/rules shall exist in CAN dir
- Only include relevant TMaps in CAN RML files
- Not the entire RML file is needed if a TMap is referenced (via parentTriplesMap)
- AcquiringCPB, Procedure(Term) not required as referenced from irrelevant TMaps
- Only TMaps related to NoticeResult shall be in LotGroup(AwardInformation)
- Entire Lot.rml.ttl is unused so remove it

Method: Search for NoticeResult in editor to see highlighted areas and
delete the rest.

TODO:

Restructure this to be a `mappings-can` folder (along with script
updates to work with it) which will contain only CAN-specific mappings,
so that the transformation can used both common and specific mappings
(along with versions), e.g.:

```sh
rmlmapper -m mappings/* mappings-can/* mappings-1.x/*
```

Packaging commands will also have to be updated to distribute the CAN
files together with the versioned files.