Add XPath similarity check to SDK change analysis scripts
Use a Python fuzzy string similarity function (partial ratio, based on a
distance metric and/or overlap ratio) to compare XPath changes.

The underlying metric(s) appear to work well for XPath strings, as two
(absolute) XPaths with only stylistic changes often have well above 90%
similarity (typically 99% but usually > 95%).

Some true changes (e.g. simple changes of listnames or
loosening/tightening of predicates/conditions) fall within the 90-95%
range, so those have to be carefully inspected.

The similarity score is currently not saved anywhere but an idea is to
prepend them to the preamble of each diff.