Add support for getting Abs XPath similarity statistics
Requires the `jq` command-line tool.

The idea is to generate diffs for a given versioned CM and get some
numbers out of that. However, theoretically, the diffs can be generated
for all versions/packages, where each subsequent run will overwrite
diffs generated in the previous one, if there is a matching field/node
and version range (the diff file is named as such).

P.S: The mode is not reported explicitly but is represented by the first
row of the frequency table, which is sorted in reverse order.