More useful Abs XPath similarity statistics
- Grouped by score
- CSV-style table

Run `analysis_scripts/analyse_versions.sh` from the project root,
against a master conceptual mapping (CM) XLSX file filtered to CN or CAN
mappings.

Otherwise, running against versioned packages' CMs would require
repeating the runs for all the versions and producing redundant diffs
(e.g. the version range 1.3-1.6 covers 4 versioned packages).

To reiterate, the diff scripts do a three-way comparison: a given min
version against a reference (1.9.1), optionally a given max against the
same reference, and if a max is given, the min against the max.

The third diff is not usually meaningful if the inputs are indeed
interpreted as min and max, as it would then compare a range that is
expected to _not_ have any functional changes.

However, it is good to be sure and look at the range diff at the same
time as the reference version. When two versions do need to be compared,
please interpret the min and max as old and new, and see the third diff.

Caveat: The version 1.5.12 is not valid -- we should probably reset it
to 1.5 in the batch analysis, as the CM as of now would contain 1.5.12
explicitly.