Initial RML pre-test for suspicious relationship mappings
This provides a list of predicate mappings involving a parentTriplesMap
that may be duplicates or wrongly placed. Possibility for many false
positives, especially for high-cardinality properties (but some of those
can be and are currently filtered out). However, good for a first
subjective evaluation.

TODO: Requires running the query against ALL RML files to yield
meaningful results.