Link from ResultNotice to AwardDecision
The AwardDecision instantiation is predicated upon the existence of an
element (AwardDate) which is in a repeatable XPath iterator
(SettledContract). Therefore the subject TMap dedicated to this linkage
needs to be conditioned slightly differently (by restricting to an
indexed XPath lookup).

However the connection itself need not be conditioned on the same
element, but can have another condition which is expected to exist at
the level of that connection (LotResult), although it is debatable
whether this is required at all.

Co-authored-by: schivmeister <schiv@archlinux.org>