Attempt to capture domain logic of OPT-100-Contract et al
According to https://github.com/OP-TED/ted-rdf-mapping-eforms/issues/7
and further feedback clarification with the client, the following logic
has been established:

- If BT-768-Contract is true, then the current Contract is of type
PurchaseContract.

- OPT-100-Contract refers to the (Result) Notice of the Framework
Agreement "under which the PurchaseContract is awarded".

In both cases the current Contract becomes a PurchaseContract, however
in the former the FrameworkAgreement contract is from another Notice
(this aspect is not yet implemented), and that Contract itself is not
known.

Additionally, from the perspective of the creation of the current
contract (ND-SettledContract), if BT-765-Lot has a value other than
`none`, then it implies the current Contract is a FrameworkAgreement.

To reiterate: if the current contract is not a FrameworkAgreement, then
it _can_ be a PurchaseContract "under" another FrameworkAgreement, which
is not known (except the Notice).

TODO: Proper instantiation of the "other" FrameworkAgreement Notice
along with proxy FrameworkAgreement instance.

Co-authored-by: schivmeister <schiv@archlinux.org>