Move common ProcurementProcessInformation back to CN
This was not required as we do not yet split up CN and common mappings.
What is there in `mappings-common` is truly shared/generic, i.e. the
file there should _not_ exist with the same name in either `mappings` or
`mappings-can`.

This actually caused an issue whereby in the packages the file was
replaced by whichever the copy process found last, as both the CN and
the common one was named the same.