Optimize subject reference condition for AwardDecision
Instead of starting with the conditional, put the base IRI construction
first, and start the conditional after that. Furthermore, we do not need
to hash the second and third cases, which removes the need for repeating
the URL-encoding segment and therefore also the remote hashing for those
cases, resulting in a shorter conditional XPath reference.