Fix Tender->Subcontractor connections latent bug
Tender-[specifiesSubcontractors]->Subcontractor

The Subcontractor was linked to only one (or the first) Tender due to an
incorrect template vs. iterator and joinCondition issue. This resolves
it such that all Tenders citing the Tenderer associated with the
Subcontractor are the ones to which the Subcontractor is linked.

A modified version of the notice 2024-OJS072-00215550 is used to test
this, as it would expose the underlying issue if the original incorrect
mapping were to exist.

Co-authored-by: csnyulas <csongor.nyulas@gmail.com>