Add packaging support for all-attributes testing
This is a superset of multilingual testing, i.e. it incorporates all
data (as in maximal) but retains the attributes-only conceptual mapping
(CM).

For tooling/testing where the intersection of the CM and SDK is taken,
note that any version < 1.9 requires inclusion of additional SDKs, at
least 1.9, as that's when most attributes were introduced (with the
exception of `-Currency`, which can be seen in 1.8).