Port orphans test as new dangling individuals validation, w/ data
Introduce the notion of an "exist" SELECT query and a counterpart
"absent" ASK query, which is a negated version of the former. Having a
negated ASK query facilitates reporting in existing tooling such as
Mapping Workbench (MWB), which flags `No` responses as errors/invalids.

Introduce also the notion of output validation test data, by repurposing
the root `test_data` folder and including RDF (not XML) data that can be
used to test these validations. They are suitable for unit testing, if
ever desired or required.

The data is deliberately _not_ reduced in size (a copy of existing CN
output), as this helps to check the real performance of a given
validation. **The queries also do not take into consideration existence
of data other than the notices in the same graph/database.**

Note that, specifically for this validation, exceptions are _not_
filtered out (like they are in the related Make testing target
`test_output`) as it is early stages for such validation in this
project, and it will be useful to observe the results for some time.