Add support for packaging minimal variant by reduction
Before this change the minimal package would be generated by doing
nothing, i.e. expecting that only one reference data sample exists in
the tracked packages or copying any and all such data over. With this
change, the minimal Make packaging targets for all current profiles (CN,
CAN) will remove all files and folders (for output) _not_ matching the
reference sample (currently the `*24_maximal.xml`).

A flag for this has been added, in the case that the tracked packages
come with the whole set of example test data from the eForms SDK. This
is turned on by default from now.