Add PlannedProcurementPart Place of Performance mappings
- BT-5101(a|b|c)-Part
- BT-5131-Part
- BT-5121-Part
- BT-5071-Part
- BT-5141-Part

Where the iterator should be for the ContractTerm and the Location vs.
Address needs to be revisited and confirmed.

Additionally, the condition for instantiating the Location/Address needs
to be rethought as checking for the country code catches also the broad
place of performance. Without it in the condition, however, those with
only the country code will not be considered at all and will miss an
instance, resulting in wrong output.

Co-authored-by: Jana <ahmadjan@fel.cvut.cz>