Add PlaceOfPerformance-related mappings, caveat emptor
BT-23-Part
BT-531-Part
BT-726-Part
BT-728-Part

WARNING: The ContracTerms could not be kept separate, as there is a
difficult situation with the iterators and template paths because
related predicates' anchoring points differ (RealizedLocation vs.
ProcurementProjectLot). Therefore, there are multiple values on the
additional information (one should have been for the specific place, and
the other for the broad). This has to be revisited.

Co-authored-by: Jana <ahmadjan@fel.cvut.cz>