Add conditional absolute root PIN mapping, caveat emptor
This is simply adding the mapping for the sake of the condition, but
does not make sense in that it has no predicates specific to the
condition (that it's a PIN notice).

This merely exists to declare an `epo-not:PlanningNotice` subject,
however the other mappings' subjects are not predicated upon such
conditions, in turn making this further pointless.

We should either add conditions everywhere else where predicates are
only supposed to be for PIN (and would be erroneous if instantiated in
other notices), or skip this conditional subject-only mapping entirely.

Similarly, the other absolute root mappings for CN and CAN should be
revisited.