Add CAN-modif hacky support to tx_source script
Add option handling and quick hack to replace any `-modif` suffix when
looking for the mappings folder, because we don't want to introduce for
now yet another folder, if the root element of such notices is still
`/ContractAwardNotice`, which is just CAN for us.

If we're happy with this, we need to update the other tx scripts as
well.