Split up mappings for BT-125(i)-Part and BT-1251-Part
There is a subtle difference which we happened to observe: the referred
notice cannot be interpreted as a PlanningNotice until and unless it has
a planning-related property, which is announcesPlannedProcurementPart.
From that point on it becomes a PlanningNotice.

We split up the refersToPrevious along the lines of
PlanningNotice->Notice and PlanningNotice->PlanningNotice. The Notice
instance should therefore have both types, and that's the only change
this produces in the output.

This is mainly a reorganization of technical mapping for consistency
with the conceptual mapping.