Add BT-717-Lot versioned mappings to cover v1.13+
Another case of a more restrictive XPath and therefore not likely to
have effect on most notices. The field itself appears from v1.9 onwards,
but changes in v1.13, along with the introduction of/in notice subtypes
E1 and E2.

We introduce also from here on the notion of applicability from a "v0",
meaning, a mapping is applicable from any point in the past, as it is
for a field which was introduced in an SDK version later than inception.