Add BT-745-Lot lang TMAp versioned mappings to cover v1.12+
The XPath became less restrictive and therefore is very unlikely to
affect any past versions of notices. This is the first time we're having
to version a language-only TriplesMap, which was a workaround to allow
RMLMapper to pick up repeating elements for the language (so the
iterator XPath includes the child).

It appears however our original mapping had the later, less restrictive
XPath, so it turns out we were always going to be covered. However, for
consistency, we correct the XPath of the older version (to be
restrictive).