Add BT-531-Part version annotation, caveat emptor
The XPath change from v1.8+ appears superficial but using it as an
iterator is not backwards- or forwards-compatible, while the older one
is. The predicate filter is shifted one level up.

This likely means we would have to add additional guards, but it is more
reasonable to request for the eForms SDK to revert the filtering
approach, as it caters to more cases/versions.