Add BT-735-LotResult versioned mapping, refactor VehicleInformation
Our versioned mapping technique is to extract out the relevant predicate
mapping into the relevant version-ranged file, so that the right XPaths,
usually in the iterator, can be provided in each case.

However, if there are other mappings that are not involved
(unversioned), they are kept in the base, common file with the original
TMap (including its subjectMap). This has undefined RML transformation
behaviour, as there would then be more than one iterator variant for the
same TMap.

Fortunately, there doesn't appear to be any adverse effect, as proven by
the (lack of) output changes from these mappings. Perhaps, something to
bring up with the RML and/or RMLMapper developers (what the specs say
vs. what is implemented)?

Regardless, we choose to reproduce the mappings in versioned files
wholesale, which is a safer approach. In this case, the base file is no
longer required. A TMap related to LotAwardOutcome is moved to its
rightful place. The GreenProcurement is kept in its own file so that
these versioned rules can refer to it.