Add BT-735-LotResult versioned mapping, caveat emptor
Our versioned mapping technique is to extract out the relevant predicate
mapping into the relevant version-ranged file, so that the right XPaths,
usually in the iterator, can be provided in each case.

However, if there are other mappings that are not involved, they are
kept in the base, common file with the original TMap. This has undefined
RML transformation behaviour, as there would then be more than one
iterator variant for the same TMap.

Forunately, there doesn't appear to be any adverse effect, as proven by
the (lack of) output changes from these mappings. Perhaps, something to
bring up with the RML and/or RMLMapper developers (what the specs say
vs. what is implemented)?