Update langMap TMaps to prevent RMLMapper errors when lang missing
When there is a language missing from an element but it is mapped as a
dedicated languageMap TriplesMap (with a custom function), the RML
processor would come across the element but not the attribute referenced
in the function, triggering the mapping but failing the function.

This behaviour can be observed with the latest patch version of the
**cn_24_multilingual** example of SDK v1.3, where it also came to our
attention that values that were expected to be language-tagged are
disappearing when their language tags are removed, but values kept
(without a language). That is another, bigger issue, though.

With regards to preventing the processing error (RMLMapper throws a
`searchString` error) a search and replace of `rml:reference "."` solved
it for the most part (except in one case where this was not in a
languageMap-specific TMap, but related to the Identifier of a cardinal
entity), with a few having conditions in the reference, which regex
capture groups could handle:

- **find:** `rml:reference "if(.*) then . else null"`
- **replace:** `rml:reference "if((exists(@languageID)) and $1) then . else null"`