Fix empty Reviewer instances, optimize BT-807-Review mapping
In some notices where the AppealInformation does not include a Party,
for e.g. `230832-2025.xml` (CAN v1.13), there were empty `epo:Reviewer`
instances. This is because we missed a condition in the referenced TMap
for the `epo:announcesRole`. However we can get rid of that and instead
just directly refer to the primary TMap generating the reviewer.