Fix VehicleInformation instantiation - runts and multiple values
Affects OPT-156-LotResult and related fields BT-723-LotResult,
BT-735-LotResult, across all lower and upper versions (v1.10 and below,
v1.13).

- Multiple values as exhibited by CAN v1.10 `481682-2024.xml`
- Runt (empty) instances as exhibited by CAN v1.13 SDK example

The fixes are many-fold, summarized as:

- Fixing of a missed iterator update in the v1.13 mapping
- Conditions for detecting more than 1 ProcurementDetails
- Condition for detecting if the category code exists
- Optimization of TMap referencing (refer directly for seeAlso)
- A second referencing LotAwardOutcome TMap for ProcurementDetails
- Some whitespace cleanup

This solves the multiple values situation by creating separate
VehicleInformation instances for each ProcurementDetails element. When
there is one, a single instance captures all information. This is
necessary as there is no way to otherwise capture all the information
given the eForms structure, as there is information "outside" the
ProcurementDetails.

Empty instances in v1.13 was namely due to a forgotten iterator update
in the corresponding versioned mapping, in the first TMap copied.