Add support for BT-1501(p)-Contract through BT-1501(s)-Contract
The former is essentially a replacement for the latter from SDK v1.11
onwards. Lifting the upper limit from the versioned mappings enables
`epo:relatesToEFormSectionIdentifier` to activate for both.

Testable with 447107-2024.xml.

WARNING: Mappings for an upper limit v1.10 stays in these versioned
files for the case of `epo:amendsContract`, where v1.11 introduces
BT-1501(c) at a different iterator for that purpose. It did not make
sense to have yet another version-ranged file just for this.